requirements.txt
streamlit==1.28.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
yfinance==0.2.18
akshare
openai==1.3.0
python-dotenv==1.0.0
pytz
ta==0.10.2
reportlab
weasyprint
schedule